requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-seed==0.3.1